[1] Renamed generic utils to specialized versions
It is required for follow up commits
